Refactored graphics, illo, and cartoons.jsx files and moved them into the [slug].jsx file.
